Add option to define aliases for tokens
This patch adds an option called `alias`, which allows the
definition of additional CSS classes for token names. This can be
useful, to combine the styling of a well known token, which is already
supported by most of the themes, with a semantically correct token name.
The option can be set to a string literal or an array of string
literals. In the following example the token name `latex-equation` is
not supported by any theme, but it will be highlighted the same as a
string.

'latex-equation': {
	pattern: /\$(\\?.)*?\$/g,
	alias: 'string'
}