Cleanup normalize-whitespace and improve keep-markup integration
This patch removes the misleading references to the parse-settings
plugin from the normalize-whitespace documentation and adds a
description for all supported operations.

The keep-markup plugin sets the property Prism.plugins.KeepMarkup
to allow the normalize-whitespace plugin to optimize its
operation.

Furthermore, the normalize-whitespace plugin makes sure, that it
is the first plugin to run on a particular code block.